refactor: 일관성을 맞추기 위해 lib 유틸리티 경로를 src/lib으로 변경
