+++ README.md
@@ -156,14 +156,16 @@ npm run dev
 - 다음 주 계획: 완성된 백엔드와 연동하여 결과 표시 및 디버깅
 
 ### Week 5 (26.01.19 ~ 26.01.25)
-- 작업 내역: 백엔드와의 통신 구현
-- AI 활용: 
-- 완료 기능:
-- 다음 주 계획:
+- 작업 내역: 프론트엔드-백엔드 간 API 통신 규격(Spec) 정의 및 연동 구현
+- AI 활용: codex 이용하여 복잡한 Nested Object을 UI 전용 상태(Flat Object)로 변환하는 어댑터 패턴 코드 자동 생성
+- 완료 기능: 
+  - 프라이머 설계 요청(Request) 프로세스 구현: 입력값 → 어댑터 → API 호출 흐름 완성
+  - 결과 모달(Result Modal) 데이터 바인딩: Mock 데이터를 활용하여 캔버스 및 리스트에 분석 결과 렌더링
+- 다음 주 계획: 사용자 입력 데이터(DNA 서열)에 대한 전처리(Sanitization) 및 유효성 검증 로직 구현
 
 
 ### Week 6 (26.01.26 ~ 26.02.01)
-- 작업 내역:
-- AI 활용:
+- 작업 내역: 사용자 입력 데이터 전처리 및 유효성 검증
+- AI 활용: DNA 서열 입력을 위한 정규식(Regex) 기반 유효성 검사 및 IUPAC 모호성 코드 처리 로직 생성
 - 완료 기능:
 - 다음 주 계획:
