+++ src/lib/api/client.ts
@@ -1,0 +1,14 @@
+import axios from "axios";
+
+const baseURL = "";
+
+export const apiClient = axios.create({
+  baseURL,
+  headers: {
+    "Content-Type": "application/json",
+  },
+});
+
+export const api = apiClient;
+
+export default apiClient;
