+++ src/types/analysis.ts
@@ -1,4 +1,4 @@
-// --- Official Spec Start ---
+  // --- Official Spec Start ---
 export interface Range {
   min: number;
   max: number;
